Change file size units from SI to IEC 80000-13:2025
Increase left-padded (%S) file size string length to 5 characters

+++ df_openbsd.go
@@ -15,5 +15,5 @@ func diskFree(wd string) string {
 	}
 
 	// Available blocks * size per block = available space in bytes
-	return "df: " + humanize(int64(uint64(stat.F_bavail)*uint64(stat.F_bsize)))
+	return "df: " + humanize(uint64(stat.F_bavail)*uint64(stat.F_bsize))
 }
