+++ df_windows.go
@@ -19,5 +19,5 @@ func diskFree(wd string) string {
 		log.Printf("diskfree: %s", err)
 		return ""
 	}
-	return "df: " + humanize(int64(free))
+	return "df: " + humanize(free)
 }
